Cassandra/requirements.txt
cassandra-driver
time_uuid
rich